Safari-609.1.11.1/LayoutTests/fast/css/order-calculated-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv"></div>
<script>
description("Tests assigning a calculated value to 'order' CSS property.");

var testDiv = document.getElementById("testDiv");

shouldBeEmptyString("testDiv.style['order']");
evalAndLog("testDiv.style['order'] = 'calc(2 * 3)'");
shouldBeEqualToString("testDiv.style['order']", "6");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('order')", "6");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
