Safari-611.1.10/LayoutTests/fast/css/order-calculated-value-expected.txt

Tests assigning a calculated value to 'order' CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['order'] is ""
testDiv.style['order'] = 'calc(2 * 3)'
PASS testDiv.style['order'] is "6"
PASS window.getComputedStyle(testDiv).getPropertyValue('order') is "6"
PASS successfullyParsed is true

TEST COMPLETE
